Support using difference kind of SSTReader (#316)
