+++ proxy_components/engine_store_ffi/src/ffi/mod.rs
@@ -4,9 +4,16 @@ pub mod raftstore_proxy_engine;
 
 pub use engine_tiflash::EngineStoreConfig;
 pub use proxy_ffi::{
-    basic_ffi_impls::*, domain_impls::*, encryption_impls::*, engine_store_helper_impls::*,
-    interfaces::root::DB as interfaces_ffi, lock_cf_reader::*, raftstore_proxy, raftstore_proxy::*,
-    raftstore_proxy_helper_impls::*, read_index_helper, sst_reader_impls::*,
+    basic_ffi_impls::*,
+    domain_impls::*,
+    encryption_impls::*,
+    engine_store_helper_impls::*,
+    interfaces::root::DB as interfaces_ffi,
+    raftstore_proxy,
+    raftstore_proxy::*,
+    raftstore_proxy_helper_impls::*,
+    read_index_helper,
+    snapshot_reader_impls::{lock_cf_reader::*, sst_reader_dispatcher::*},
 };
 
 pub use self::raftstore_proxy_engine::*;
