+++ proxy_components/proxy_ffi/src/lib.rs
@@ -16,18 +16,17 @@ pub mod context_impls;
 pub mod encryption_impls;
 // FFI directly related with EngineStoreServerHelper.
 pub mod engine_store_helper_impls;
-pub mod lock_cf_reader;
 // FFI directly related with RaftStoreProxyFFIHelper.
 pub mod raftstore_proxy;
 pub mod raftstore_proxy_helper_impls;
 pub mod read_index_helper;
-pub mod sst_reader_impls;
+// FFI releated with reading from SST/RocksDB files.
+pub mod snapshot_reader_impls;
 pub mod utils;
 
 pub use self::{
     basic_ffi_impls::*, domain_impls::*, encryption_impls::*, engine_store_helper_impls::*,
-    interfaces::root::DB as interfaces_ffi, lock_cf_reader::*, raftstore_proxy::*,
-    raftstore_proxy_helper_impls::*, sst_reader_impls::*,
+    interfaces::root::DB as interfaces_ffi, raftstore_proxy::*, raftstore_proxy_helper_impls::*,
 };
 
 #[allow(clippy::wrong_self_convention)]
