+++ proxy_components/proxy_ffi/src/snapshot_reader_impls/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2023 TiKV Project Authors. Licensed under Apache-2.0.
+
+pub mod lock_cf_reader;
+pub mod sst_file_reader;
+pub mod sst_reader_dispatcher;
+pub mod tablet_reader;
+
+pub use lock_cf_reader::*;
+pub use sst_reader_dispatcher::*;
