+++ raftstore-proxy/ffi/src/RaftStoreProxyFFI/@version
@@ -1,3 +1,3 @@
 #pragma once
 #include <cstdint>
-namespace DB { constexpr uint64_t RAFT_STORE_PROXY_VERSION = 4990756589462826693ull; }
+namespace DB { constexpr uint64_t RAFT_STORE_PROXY_VERSION = 15438096667077497650ull; }
