+++ raftstore-proxy/ffi/src/RaftStoreProxyFFI/ProxyFFI.h
@@ -154,6 +154,7 @@ struct RaftStoreProxyPtr {
 
 struct SSTReaderPtr {
   RawVoidPtr inner;
+  uint64_t kind;
 };
 
 struct SSTReaderInterfaces {
